locale.sh: remove leftovers
We were in certain situations still writing to locale.sh. While it would
not have caused any harm, it was unintended and now removed.

This change means that locale.sh will no longer load the consolemap. This
would only be useful if it had been reset accidentally, and logging in again
would then fix a broken console. The issue is described in FS#18759. It is
not clear to me if this is still an issue, or if it solved by our new
utf8/non-utf8 handling.

Either way, resetting the consolemap unconditionaly does not make much sense
in case the user has changed it on purpose.

It might make sense to factor out the vconsole initialization into a
"setup/repair my console" script, similar to systemd's
/lib/systemd/systemd-vconsole-setup, but that would have to wait for a
future release.

Reported-by: Dave Reisner <d@falconindy.com>
Cc: Роман Кирилич <roman.kyrylych@gmail.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>