fsck: skip remounting / fsck'ing the root fs
In case the initramfs already fsck'ed and mounted the root fs ro, we don't remount ro and
fsck.

We still remount the rootfs rw unconditionally at the end. We should probably stop doing
this and require the user to specify rw/ro explicitly in fstab.

We assume that the initramfs will also fsck and mount /usr, so we do not fsck any
mounted fs (not restricted to the rootfs).

Signed-off-by: Tom Gundersen <teg@jklm.no>