mount: no longer pass '-n' to mount
This was needed when /etc/mtab was a regular file in order to avoid writing to it
when /etc/ was mounted ro. However, now it is not needed.

Passing -n should have been a noop, except for the NILFS mount helper. It will only
start the daemon if the -n flag is not passed.

This fixes FS#22523.

Tested-by: Alexander Lam <lambchop468@gmail.com>
Signed-off-by: Tom Gundersen <teg@jklm.no>