udev: don't override settle timeout
The Arch default was 30 but upstream default was 120. This caused problems when
udev-177-3 introduced a 30 second timeout for broken kernel drivers.

We don't want to do this kind of fiddling in initscripts, if there is a need for
overriding the timeout this should be added upstream to udev.conf. My feeling is
that this is a bad idea, as there are now at least three different kinds of timeouts
at play in udev, and if they are set to something incompatible it will cause unneccesary
problems.

Signed-off-by: Tom Gundersen <teg@jklm.no>