Remove NEED_ROOT crap
This patch revert patch 042d197b4d989ec64.

NEED_ROOT was introduced to fix bug FS#24095. But in fact it doesn't solve it
because nobody use it. The idea was to allow initscripts developper to have
a check runned by every scripts which tell if scripts need to be run as root
(most of them) or not (and the script check itself which part of him have to be
run as root). All this to display a cute error message.
I think this complexity is superfluous and let part of initscript will fail if
rights are not enough.
I must confess that we should have marked this bug as wontfix rather than do that.

Signed-off-by: Sébastien Luttringer <seblu@seblu.net>